+++ data/server.go
@@ -1,0 +1,7 @@
+package data
+
+// Server represents a server
+type Server struct {
+	Prefix string
+}
+
