+++ data/user.go
@@ -1,4 +1,4 @@
-package main
+package data
 
 import (
 	"github.com/bwmarrin/discordgo"
