+++ room/npcs.go
